+++ zephyr/src/backend/minimal/spotflow_minimal_backend.h
@@ -1,0 +1,7 @@
+#ifndef SPOTFLOW_MINIMAL_BACKEND_H_
+#define SPOTFLOW_MINIMAL_BACKEND_H_
+
+void spotflow_minimal_backend_init(void);
+int spotflow_minimal_backend_process_one(void);
+
+#endif /* SPOTFLOW_MINIMAL_BACKEND_H_ */